# apps/gateful/openapi/gateful.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anticapture Gateful REST API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "DelegationsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DelegationItem"
            }
          },
          "totalCount": {
            "type": "integer"
          }
        },
        "required": ["items", "totalCount"],
        "description": "Paginated historical delegations response."
      },
...
      "OrderDirection": {
        "type": "string",
        "enum": ["asc", "desc"],
        "default": "desc",
        "description": "Sort direction for ordered query results."
      },
      "DelegatorsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DelegatorItem"
            }
          },
          "totalCount": {
            "type": "integer"
          }
        },
        "required": ["items", "totalCount"],
        "description": "Paginated delegators for a delegate address."
      },
...
      "ProposalActivityResponse": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "format": "ethereum-address",
            "description": "Delegate address."
          },
          "totalProposals": {
            "type": "integer",
            "description": "Total proposals reviewed in the dataset."
          },
          "votedProposals": {
            "type": "integer",
            "description": "Number of proposals the delegate voted on."
          },
          "neverVoted": {
            "type": "boolean",
            "description": "Whether the delegate never cast a vote."
          },
          "winRate": {
            "type": "number",
            "description": "Share of proposals where the delegate sided with outcome."
          },
          "yesRate": {
            "type": "number",
            "description": "Share of delegate votes cast in support."
          },
          "avgTimeBeforeEnd": {
            "type": "number",
            "description": "Average seconds between the delegate vote and proposal end time."
          },
          "proposals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalActivityItem"
            }
          }
        },
        "required": [
          "address",
          "totalProposals",
          "votedProposals",
          "neverVoted",
          "winRate",
          "yesRate",
          "avgTimeBeforeEnd",
          "proposals"
        ],
        "description": "Delegate proposal activity metrics and proposal-by-proposal history."
      },
...
      "OnchainProposalStatusList": {
        "type": "array",
        "items": {
          "type": "string",
          "enum": [
            "PENDING",
            "ACTIVE",
            "CANCELED",
            "DEFEATED",
            "SUCCEEDED",
            "QUEUED",
            "EXPIRED",
            "PENDING_EXECUTION",
            "EXECUTED",
            "VETOED",
            "NO_QUORUM"
          ]
        },
        "description": "Proposal status filter. Pass repeated query params or a comma-delimited list.",
        "example": ["ACTIVE"]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/{dao}/accounts/{address}/delegations/historical": {
      "get": {
        "operationId": "historicalDelegations",
        "summary": "Get historical delegations",
        "description": "Get historical delegations for an account, with optional filtering and sorting",
        "tags": ["delegations", "skip-pagination"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "address",
            "in": "path"
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false,
            "name": "delegateAddressIn",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Number of records to skip before collecting results.",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip before collecting results.",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Maximum number of records to return.",
              "example": 10
            },
            "required": false,
            "description": "Maximum number of records to return.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "fromValue",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "toValue",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            },
            "required": false,
            "description": "Sort direction for ordered query results.",
            "name": "orderDirection",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns historical delegations for an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationsResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "aave",
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
    "/{dao}/accounts/{address}/delegations": {
      "get": {
        "operationId": "delegations",
        "summary": "Get delegations",
        "description": "Get current delegations for an account",
        "tags": ["delegations"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "address",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns delegations for an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationsResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "aave",
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
    "/{dao}/accounts/{address}/delegators": {
      "get": {
        "operationId": "delegators",
        "summary": "Get delegators",
        "description": "Get current delegators of an account with voting power",
        "tags": ["delegations", "skip-pagination"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "address",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Number of records to skip before collecting results.",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip before collecting results.",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Maximum number of records to return.",
              "example": 10
            },
            "required": false,
            "description": "Maximum number of records to return.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["amount", "timestamp"],
              "default": "amount"
            },
            "required": false,
            "name": "orderBy",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            },
            "required": false,
            "description": "Sort direction for ordered query results.",
            "name": "orderDirection",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns current delegators for an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegatorsResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "aave",
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
...
    "/{dao}/proposals-activity": {
      "get": {
        "operationId": "proposalsActivity",
        "summary": "Get proposals activity for delegate",
        "description": "Returns proposal activity data including voting history, win rates, and detailed proposal information for the specified delegate within the given time window",
        "tags": ["proposals", "skip-pagination"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Delegate address whose proposal activity is being queried.",
              "example": "0x1111111111111111111111111111111111111111"
            },
            "required": true,
            "description": "Delegate address whose proposal activity is being queried.",
            "name": "address",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Lower bound for proposal timestamps, in Unix seconds.",
              "example": 1700000000
            },
            "required": false,
            "description": "Lower bound for proposal timestamps, in Unix seconds.",
            "name": "fromDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Number of proposal activity records to skip.",
              "example": 0
            },
            "required": false,
            "description": "Number of proposal activity records to skip.",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Maximum number of proposal activity records to return.",
              "example": 10
            },
            "required": false,
            "description": "Maximum number of proposal activity records to return.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["timestamp", "votingPower", "voteTiming"],
              "default": "timestamp",
              "description": "Field used to sort proposal activity results.",
              "example": "timestamp"
            },
            "required": false,
            "description": "Field used to sort proposal activity results.",
            "name": "orderBy",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            },
            "required": false,
            "description": "Sort direction for ordered query results.",
            "name": "orderDirection",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["yes", "no", "abstain", "no_vote"],
              "description": "Optional vote filter. Use yes, no, abstain, or no-vote to narrow the result set.",
              "example": "yes"
            },
            "required": false,
            "description": "Optional vote filter. Use yes, no, abstain, or no-vote to narrow the result set.",
            "name": "userVoteFilter",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved proposals activity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalActivityResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
    "/{dao}/proposals": {
      "get": {
        "operationId": "proposals",
        "summary": "Get proposals",
        "description": "Returns a list of proposal",
        "tags": ["proposals", "skip-pagination"],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "description": "Number of records to skip before collecting results.",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip before collecting results.",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Maximum number of records to return.",
              "example": 10
            },
            "required": false,
            "description": "Maximum number of records to return.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            },
            "required": false,
            "description": "Sort direction for ordered query results.",
            "name": "orderDirection",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OnchainProposalStatusList"
            },
            "required": false,
            "description": "Proposal status filter. Pass repeated query params or a comma-delimited list.",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Earliest proposal timestamp, in Unix seconds.",
              "example": 1700000000
            },
            "required": false,
            "description": "Earliest proposal timestamp, in Unix seconds.",
            "name": "fromDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Latest proposal end timestamp, in Unix seconds.",
              "example": 1700000000
            },
            "required": false,
            "description": "Latest proposal end timestamp, in Unix seconds.",
            "name": "fromEndDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": true,
              "description": "Whether optimistic proposals should be included.",
              "example": false
            },
            "required": false,
            "description": "Whether optimistic proposals should be included.",
            "name": "includeOptimisticProposals",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved proposals activity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnchainProposalsResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
...
    "/{dao}/accounts/{address}/balances/historical": {
      "get": {
        "operationId": "historicalBalances",
        "summary": "Get historical token balances",
        "description": "Returns historical balance deltas for one account, enriched with the transfer that caused each change.",
        "tags": ["account-balances", "skip-pagination"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "address",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Number of historical balance rows to skip.",
              "example": 0
            },
            "required": false,
            "description": "Number of historical balance rows to skip.",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Maximum number of historical balance rows to return.",
              "example": 10
            },
            "required": false,
            "description": "Maximum number of historical balance rows to return.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["timestamp", "delta"],
              "default": "timestamp",
              "description": "Field used to sort historical balance rows.",
              "example": "timestamp"
            },
            "required": false,
            "description": "Field used to sort historical balance rows.",
            "name": "orderBy",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            },
            "required": false,
            "description": "Sort direction for ordered query results.",
            "name": "orderDirection",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Inclusive lower bound for historical balance timestamps, in Unix seconds.",
              "example": 1700000000
            },
            "required": false,
            "description": "Inclusive lower bound for historical balance timestamps, in Unix seconds.",
            "name": "fromDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "Inclusive upper bound for historical balance timestamps, in Unix seconds.",
              "example": 1700000000
            },
            "required": false,
            "description": "Inclusive upper bound for historical balance timestamps, in Unix seconds.",
            "name": "toDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Minimum balance delta encoded as a decimal string."
            },
            "required": false,
            "description": "Minimum balance delta encoded as a decimal string.",
            "name": "fromValue",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Maximum balance delta encoded as a decimal string."
            },
            "required": false,
            "description": "Maximum balance delta encoded as a decimal string.",
            "name": "toValue",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved historical balances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalBalancesResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "aave",
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
...
    "/{dao}/last-update": {
      "get": {
        "operationId": "lastUpdate",
        "summary": "Get the last update time",
        "tags": ["metrics"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "cost_comparison",
                "attack_profitability",
                "token_distribution"
              ],
              "description": "Chart identifier whose freshness timestamp should be returned.",
              "example": "token_distribution"
            },
            "required": true,
            "description": "Chart identifier whose freshness timestamp should be returned.",
            "name": "chart",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Last update time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LastUpdateResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    },
...
    "/{dao}/tokens/available": {
      "get": {
        "operationId": "availableTokens",
        "summary": "Get available tokens",
        "description": "Get available ERC-20 tokens for transfer in the DAO's governance chain",
        "tags": ["tokens"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns available ERC-20 tokens for the DAO's chain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaoTokensResponse"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "dao",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "aave",
              "comp",
              "ens",
              "fluid",
              "gtc",
              "lil_nouns",
              "nouns",
              "obol",
              "scr",
              "shu",
              "uni"
            ]
          },
          "description": "DAO identifier"
        }
      ]
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}